Refactor navigator assembly orchestration and telegram adapters
